Convert hook to HoC
